requirements.txt
RPi.GPIO==0.7.0
pyzmq==21.0.1